refactor: post 단건 조회 -> QueryDSL로 Member와 Cat 정보 포함 resposne

+++ src/main/java/com/ogu/soonnyang/domain/post/dto/CreatePostRequest.java
@@ -11,9 +11,9 @@ public class CreatePostRequest {
 
     private Long catId;
 
-    private float latitude;
+    private Double latitude;
 
-    private float longitude;
+    private Double longitude;
 
     private String content;
