+++ src/main/java/com/ogu/soonnyang/domain/post/entity/Post.java
@@ -20,9 +20,9 @@ public class Post extends BaseTimeEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long postId;
 
-    private Float latitude;  // 위도
+    private Double latitude;  // 위도
 
-    private Float longitude;  // 경도
+    private Double longitude;  // 경도
 
     private String content;
